Time: 18 ms (25.89%), Space: 87.3 MB (58.56%) - LeetHub

+++ 0111-minimum-depth-of-binary-tree/0111-minimum-depth-of-binary-tree.java
@@ -25,6 +25,6 @@ class Solution {
         
 		int left = minDepth(node.left);
 		int right = minDepth(node.right);
-		return (int)Math.min(left, right) + 1;
+		return Math.min(left, right) + 1;
     }
 }
